tests refactored + protected to private
