+++ src/main/java/in/zapr/druid/druidry/dimension/ListFilteredDimension.java
@@ -14,9 +14,9 @@ import java.util.List;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @EqualsAndHashCode(callSuper = true)
 public class ListFilteredDimension extends FilteredDimension {
-    protected List<String> values;
+    private List<String> values;
     @JsonProperty(value = "isWhitelist")
-    protected Boolean whitelist;
+    private Boolean whitelist;
 
     @Builder
     public ListFilteredDimension(@NonNull DimensionSpec dimensionSpec, @NonNull List<String> values, Boolean whitelist) {
