+++ src/test/java/in/zapr/druid/druidry/dimension/ListFilteredDimensionTest.java
@@ -10,100 +10,12 @@ import org.testng.annotations.Test;
 
 import java.util.Arrays;
 
-public class FilteredDimensionTest {
+public class ListFilteredDimensionTest {
     private static ObjectMapper objectMapper;
 
     @BeforeClass
     public void init() {
         objectMapper = new ObjectMapper();
-    }
-
-
-    @Test
-    public void testSampleRegexFilteredDimension() throws JsonProcessingException, JSONException {
-        DimensionSpec dimensionSpec = DefaultDimension.builder()
-                .dimension("system_label_values")
-                .outputName("system_label_values")
-                .build();
-
-        RegexFilteredDimension regexFilteredDimension = RegexFilteredDimension.builder()
-                .dimensionSpec(dimensionSpec)
-                .pattern("compute.googleapis.com/cores`.*")
-                .build();
-
-        String jsonOutput = objectMapper.writeValueAsString(regexFilteredDimension);
-        String expectedJSONString = "{\n" +
-                "      \"type\": \"regexFiltered\",\n" +
-                "      \"delegate\": {\n" +
-                "        \"type\": \"default\",\n" +
-                "        \"dimension\": \"system_label_values\",\n" +
-                "        \"outputName\": \"system_label_values\"\n" +
-                "      },\n" +
-                "      \"pattern\": \"compute.googleapis.com/cores`.*\"\n" +
-                "    }";
-        JSONAssert.assertEquals(expectedJSONString, jsonOutput, JSONCompareMode.NON_EXTENSIBLE);
-    }
-
-    @Test(expectedExceptions = NullPointerException.class)
-    public void testSampleRegexFilteredDimensionWithNullDimensionSpec() throws JsonProcessingException {
-        RegexFilteredDimension regexFilteredDimension = RegexFilteredDimension.builder()
-                .pattern("compute.googleapis.com/cores`.*")
-                .build();
-    }
-
-    @Test(expectedExceptions = NullPointerException.class)
-    public void testSampleRegexFilteredDimensionWithNullRegex() throws JsonProcessingException {
-        DimensionSpec dimensionSpec = DefaultDimension.builder()
-                .dimension("system_label_values")
-                .outputName("system_label_values")
-                .build();
-        RegexFilteredDimension regexFilteredDimension = RegexFilteredDimension.builder()
-                .dimensionSpec(dimensionSpec)
-                .build();
-    }
-
-    @Test
-    public void testPreFixFilteredDimension() throws JsonProcessingException, JSONException {
-        DimensionSpec dimensionSpec = DefaultDimension.builder()
-                .dimension("system_label_values")
-                .outputName("system_label_values")
-                .build();
-        PrefixFilteredDimension prefixFilteredDimension = PrefixFilteredDimension.builder()
-                .dimensionSpec(dimensionSpec)
-                .prefix("compute.googleapis.com/cores")
-                .build();
-
-        String jsonOutput = objectMapper.writeValueAsString(prefixFilteredDimension);
-        String expectedJSONString = "{\n" +
-                "      \"type\": \"prefixFiltered\",\n" +
-                "      \"delegate\": {\n" +
-                "        \"type\": \"default\",\n" +
-                "        \"dimension\": \"system_label_values\",\n" +
-                "        \"outputName\": \"system_label_values\"\n" +
-                "      },\n" +
-                "      \"prefix\": \"compute.googleapis.com/cores\"\n" +
-                "    }";
-        JSONAssert.assertEquals(expectedJSONString, jsonOutput, JSONCompareMode.NON_EXTENSIBLE);
-    }
-
-    @Test(expectedExceptions = NullPointerException.class)
-    public void testPreFixFilteredDimensionWithNullDimensionSpec() throws JsonProcessingException {
-        PrefixFilteredDimension prefixFilteredDimension = PrefixFilteredDimension.builder()
-                .prefix("compute.googleapis.com/cores")
-                .build();
-
-    }
-
-    @Test(expectedExceptions = NullPointerException.class)
-    public void testPreFixFilteredDimensionWithNullPrefix() throws JsonProcessingException {
-        DimensionSpec dimensionSpec = DefaultDimension.builder()
-                .dimension("system_label_values")
-                .outputName("system_label_values")
-                .build();
-
-        PrefixFilteredDimension prefixFilteredDimension = PrefixFilteredDimension.builder()
-                .dimensionSpec(dimensionSpec)
-                .build();
     }
 
 
@@ -204,5 +116,6 @@ public class FilteredDimensionTest {
                 .dimensionSpec(dimensionSpec)
                 .build();
     }
+
 
 }
